Refactor: Move the injection-capable channel to its own submodule
